magic_georeferencer/matchanything/inference.py
"""
MatchAnything Inference Wrapper for Magic Georeferencer

Simplified inference interface for the MatchAnything deep learning model.
"""

import numpy as np
import torch
import cv2
from pathlib import Path
from typing import Tuple, Optional
from PIL import Image


class MatchAnythingInference:
    """Simplified MatchAnything inference wrapper"""

    def __init__(self, weights_path: Path, device: str = 'auto'):
        """Initialize MatchAnything model.

        Args:
            weights_path: Path to model weights directory
            device: 'cuda', 'cpu', or 'auto' (auto-detect)
        """
        self.weights_path = Path(weights_path)

        # Auto-detect device if needed
        if device == 'auto':
            self.device = 'cuda' if torch.cuda.is_available() else 'cpu'
        else:
            self.device = device

        # Load model
        self.model = self._load_model()
        if self.model is not None:
            self.model.to(self.device)
            self.model.eval()

    def _load_model(self):
        """Load MatchAnything model from weights.

        Returns:
            Loaded model or None if loading fails

        Note: This is a placeholder implementation. The actual implementation
        will depend on the MatchAnything model architecture and API.
        """
        try:
            # TODO: Replace with actual MatchAnything loading code
            # This will depend on their model architecture

            # Placeholder: Check for common weight file names
            weight_candidates = [
                'model_checkpoint.pth',
                'model.pth',
                'checkpoint.pth',
                'weights.pth'
            ]

            weight_file = None
            for candidate in weight_candidates:
                candidate_path = self.weights_path / candidate
                if candidate_path.exists():
                    weight_file = candidate_path
                    break

            if weight_file is None:
                # Check for any .pth file
                pth_files = list(self.weights_path.glob('*.pth'))
                if pth_files:
                    weight_file = pth_files[0]
                else:
                    raise FileNotFoundError(f"No weight files found in {self.weights_path}")

            # TODO: Replace this with actual MatchAnything model loading
            # For now, return a placeholder that will be replaced
            # when integrating the actual MatchAnything code

            # Example structure (depends on actual MatchAnything API):
            # from matchanything.model import MatchAnythingModel
            # model = MatchAnythingModel.load_from_checkpoint(weight_file)

            print(f"Found weights at: {weight_file}")
            return None  # Placeholder - will be replaced with actual model

        except Exception as e:
            print(f"Error loading model: {e}")
            raise

    def match_images(
        self,
        image1: np.ndarray,
        image2: np.ndarray,
        max_keypoints: int = 2048
    ) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
        """Match two images and return corresponding keypoints.

        Args:
            image1: First image [H, W, 3] RGB numpy array
            image2: Second image [H, W, 3] RGB numpy array
            max_keypoints: Maximum number of keypoints to detect

        Returns:
            Tuple of (keypoints1, keypoints2, confidence):
            - keypoints1: [N, 2] coordinates in image1 (x, y)
            - keypoints2: [N, 2] coordinates in image2 (x, y)
            - confidence: [N] confidence score per match (0-1)
        """
        if self.model is None:
            raise RuntimeError("Model not loaded. Cannot perform matching.")

        # Preprocess images
        img1_tensor = self._preprocess_image(image1)
        img2_tensor = self._preprocess_image(image2)

        # Run inference
        with torch.no_grad():
            # TODO: Replace with actual MatchAnything inference call
            # This is a placeholder structure

            # Example (depends on actual MatchAnything API):
            # results = self.model(img1_tensor, img2_tensor, max_keypoints=max_keypoints)

            # For now, return dummy results for testing structure
            # This will be replaced when integrating actual MatchAnything
            num_matches = min(max_keypoints, 100)  # Placeholder

            keypoints1 = np.random.rand(num_matches, 2) * np.array([image1.shape[1], image1.shape[0]])
            keypoints2 = np.random.rand(num_matches, 2) * np.array([image2.shape[1], image2.shape[0]])
            confidence = np.random.rand(num_matches)

        return keypoints1, keypoints2, confidence

    def _preprocess_image(self, image: np.ndarray) -> torch.Tensor:
        """Preprocess image for MatchAnything inference.

        Args:
            image: Input image as numpy array [H, W, 3] in RGB format

        Returns:
            Preprocessed image tensor ready for model input
        """
        # TODO: Replace with actual MatchAnything preprocessing requirements

        # Common preprocessing steps (may need adjustment):
        # 1. Convert to float and normalize
        image_float = image.astype(np.float32) / 255.0

        # 2. Convert to tensor
        image_tensor = torch.from_numpy(image_float).permute(2, 0, 1)  # [C, H, W]

        # 3. Add batch dimension
        image_tensor = image_tensor.unsqueeze(0)  # [1, C, H, W]

        # 4. Move to device
        image_tensor = image_tensor.to(self.device)

        return image_tensor

    def estimate_homography(
        self,
        keypoints1: np.ndarray,
        keypoints2: np.ndarray,
        confidence: np.ndarray,
        ransac_threshold: float = 3.0
    ) -> Tuple[Optional[np.ndarray], np.ndarray]:
        """Estimate homography matrix using RANSAC.

        Args:
            keypoints1: [N, 2] keypoints from image 1
            keypoints2: [N, 2] keypoints from image 2
            confidence: [N] confidence scores
            ransac_threshold: RANSAC reprojection threshold in pixels

        Returns:
            Tuple of (H, inlier_mask):
            - H: [3, 3] homography matrix (or None if failed)
            - inlier_mask: [N] boolean mask of inliers
        """
        if len(keypoints1) < 4:
            # Need at least 4 points for homography
            return None, np.zeros(len(keypoints1), dtype=bool)

        try:
            # Use OpenCV's RANSAC homography estimation
            H, mask = cv2.findHomography(
                keypoints1,
                keypoints2,
                cv2.RANSAC,
                ransac_threshold
            )

            if H is None:
                return None, np.zeros(len(keypoints1), dtype=bool)

            inlier_mask = mask.ravel().astype(bool)
            return H, inlier_mask

        except cv2.error as e:
            print(f"Homography estimation failed: {e}")
            return None, np.zeros(len(keypoints1), dtype=bool)

    def resize_image(
        self,
        image: np.ndarray,
        target_size: int,
        maintain_aspect: bool = True
    ) -> Tuple[np.ndarray, float]:
        """Resize image to target size.

        Args:
            image: Input image [H, W, 3]
            target_size: Target size (longest edge)
            maintain_aspect: Whether to maintain aspect ratio

        Returns:
            Tuple of (resized_image, scale_factor)
        """
        h, w = image.shape[:2]

        if maintain_aspect:
            # Scale based on longest edge
            scale = target_size / max(h, w)
            new_h = int(h * scale)
            new_w = int(w * scale)
        else:
            new_h = target_size
            new_w = target_size
            scale = target_size / max(h, w)

        resized = cv2.resize(image, (new_w, new_h), interpolation=cv2.INTER_LINEAR)

        return resized, scale

    def scale_keypoints(
        self,
        keypoints: np.ndarray,
        scale: float
    ) -> np.ndarray:
        """Scale keypoint coordinates.

        Args:
            keypoints: [N, 2] keypoint coordinates
            scale: Scale factor

        Returns:
            Scaled keypoints [N, 2]
        """
        return keypoints * scale
