magic_georeferencer/matchanything/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.8.0
pillow>=9.0.0
requests>=2.28.0
